fix issues with the failed tests
