+++ test/imagekit/spec_helper.rb
@@ -10,5 +10,6 @@ require 'webmock/rspec'
 require 'carrierwave'
 require 'byebug'
 require 'active_storage'
+require "rspec/autorun"
 
 require_relative '../../lib/imagekitio'
